# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "Backend ::Deliah Restó:: App",
      "version": "1.0.0",
      "title": "Deliah Restó REST API",
      "contact": {
        "email": "carla.a.capocci@gmail.com"
      },
      "license": {
        "name": "Apache 2.0",
        "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      }
    },
    "host": "localhost:33333",
    "basePath": "/",
...
    "schemes": [
      "http"
    ],
    "paths": {
...
      "/pedidos": {
        "get": {
          "tags": [
            "Pedidos"
          ],
          "summary": "traer",
          "description": "trae todos los pedidos solo para el admin",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "header",
              "name": "Authorization",
              "description": "JWT Token",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "listado de favoritos",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Pedidos"
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "Pedidos"
          ],
          "summary": "agregar",
          "description": "Agregar un pedido",
          "parameters": [
            {
              "in": "header",
              "name": "Authorization",
              "description": "JWT Token",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/mis-pedidos": {
        "get": {
          "tags": [
            "Pedidos"
          ],
          "summary": "traer",
          "description": "trae todos los pedidos del usuario",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "header",
              "name": "Authorization",
              "description": "JWT Token",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "listado de pedidos",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Pedidos"
                }
              }
            }
          }
        }
      },
      "/pedidos/{id}": {
        "delete": {
          "tags": [
            "Pedidos"
          ],
          "summary": "eliminar",
          "description": "elimina un pedido",
          "parameters": [
            {
              "name": "id",
              "required": true,
              "in": "path",
              "type": "string",
              "description": "Id del pedido"
            },
            {
              "in": "header",
              "name": "Authorization",
              "description": "JWT Token",
              "required": true,
              "type": "string"
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/Pedidos"
              }
            }
          }
        },
        "patch": {
          "tags": [
            "Pedidos"
          ],
          "summary": "editar",
          "description": "solo el admin",
          "parameters": [
            {
              "name": "id",
              "required": true,
              "in": "path",
              "type": "string",
              "description": "Id del pedido"
            },
            {
              "in": "header",
              "name": "Authorization",
              "description": "JWT Token",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "datos",
              "description": "datos de pedido",
              "schema": {
                "type": "object",
                "properties": {
                  "fecha": {
                    "type": "string"
                  },
                  "user_id": {
                    "type": "integer"
                  },
                  "fp_id": {
                    "type": "integer"
                  },
                  "estado_id": {
                    "type": "integer"
                  }
                }
              }
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/ResponseOK"
              }
            }
          }
        }
      },
      "/pedidos/{idPedido}/platos/": {
        "post": {
          "tags": [
            "Pedidos"
          ],
          "summary": "agregar",
          "description": "Agregar un plato a un pedido",
          "parameters": [
            {
              "in": "header",
              "name": "Authorization",
              "description": "JWT Token",
              "required": true,
              "type": "string"
            },
            {
              "name": "idPedido",
              "required": true,
              "in": "path",
              "type": "string",
              "description": "Id del pedido"
            },
            {
              "in": "body",
              "name": "datos",
              "description": "datos del pedido plato",
              "schema": {
                "type": "object",
                "properties": {
                  "cantidad": {
                    "type": "integer"
                  },
                  "plato_id": {
                    "type": "integer"
                  }
                }
              }
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/ResponseOK"
              }
            }
          }
        }
      },
      "/pedidos/{idPedido}/plato/{idPedidosPlato}": {
        "patch": {
          "tags": [
            "Pedidos"
          ],
          "summary": "editar",
          "description": "solo el admin",
          "parameters": [
            {
              "name": "idPedido",
              "required": true,
              "in": "path",
              "type": "string",
              "description": "Id del pedido"
            },
            {
              "name": "idPedidosPlato",
              "required": true,
              "in": "path",
              "type": "string",
              "description": "Id del pedido plato"
            },
            {
              "in": "header",
              "name": "Authorization",
              "description": "JWT Token",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "datos",
              "description": "datos de pedido",
              "schema": {
                "type": "object",
                "properties": {
                  "plato_id": {
                    "type": "integer"
                  },
                  "precio": {
                    "type": "number"
                  },
                  "cantidad": {
                    "type": "number"
                  }
                }
              }
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/ResponseOK"
              }
            }
          }
        },
        "delete": {
          "tags": [
            "Pedidos"
          ],
          "summary": "eliminar",
          "description": "elimina un pedido plato",
          "parameters": [
            {
              "name": "idPedido",
              "required": true,
              "in": "path",
              "type": "string",
              "description": "Id del pedido"
            },
            {
              "name": "idPedidosPlato",
              "required": true,
              "in": "path",
              "type": "string",
              "description": "Id del pedido plato"
            },
            {
              "in": "header",
              "name": "Authorization",
              "description": "JWT Token",
              "required": true,
              "type": "string"
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/Pedidos"
              }
            }
          }
        }
      }
    },
    "definitions": {
...
      "ResponseOK": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          }
        }
      },
...
      "Pedidos": {
        "type": "object",
        "properties": {
          "pedido_id": {
            "type": "integer"
          },
          "fecha": {
            "type": "string"
          },
          "user_id": {
            "type": "integer"
          },
          "fp_id": {
            "type": "integer"
          },
          "estado_id": {
            "type": "integer"
          },
          "platos": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "pedidos-plato_id": {
                  "type": "integer"
                },
                "plato_id": {
                  "type": "integer"
                },
                "precio": {
                  "type": "number"
                },
                "cantidad": {
                  "type": "number"
                },
                "nombre": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
    }
  }